Connected Home Activity views to the database
This commit migrates to using the Room database. The database is
accessed via a ViewModel, which connects to the AppRepository. The
repository connects directly to the DB and the Data Access Objects
(DAO)s. The DAOs define an interface to execute queries on the database.

The dependency and disability fragments are attached to the database via
the ViewModel.

The views populate and update themselves automatically using the
LiveData retrieved from the database queries.

Users can set their dependency status and create any number of
disabilities, all of which is stored in the database. The user can
create, edit, and delete disabilities within the disabilities fragment.

TODO: Enable adding SMC ratings.
TODO: Add a Dialogue for adding dependents with birth defects.
TODO: Update the total calculation based on user data.