Added Birth Defect ListView.
Added the ListView that shows current children with birth defects.

Added the XML for the ListView item view.

Added an adapter for the above list view.

Added a enumeration class that defines the action states of Dialog windows.
States include CREATE and EDIT.

Added listeners to buttons displayed within the Birth Defect
ListView that deletes an item, or opens the Birth Defect Dialog
passing it the action and ID of the item selected.

Fixed a typo: renamed several files.

Removed unused code from the ViewModel